# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Prueba",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/auth/login": {
            "post": {
                "summary": "Obtiene el token despues de Loguearse",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "prueba@gmail.com",
                                    "password": "123456789"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Obtiene el token"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/auth/getCiudades": {
            "get": {
                "summary": "Obtiene las Ciudades de Argentina con sus provincias",
                "operationId": "7c6814de2d40c3c68a6bcf1c218355b9",
                "responses": {
                    "200": {
                        "description": "Obtiene las Ciudades de Argentina"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Introduzca su token con el siguiente formato (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "api_key_security_example": []
        }
    ]
}